Move OutsideURLMiddleware to new middleware.py.
