+++ server/devpi_server/main.py
@@ -336,9 +336,9 @@ class XOM:
     view_deriver.options = ('is_mutating',)
 
     def create_app(self):
+        from devpi_server.middleware import OutsideURLMiddleware
         from devpi_server.view_auth import DevpiSecurityPolicy
         from devpi_server.views import ContentTypePredicate
-        from devpi_server.views import OutsideURLMiddleware
         from devpi_server.views import route_url, INSTALLER_USER_AGENT
         from pkg_resources import get_distribution
         from pyramid.config import Configurator
