+++ server/devpi_server/views.py
@@ -146,26 +146,6 @@ class ContentTypePredicate(object):
         return request.content_type == self.val
 
 
-class OutsideURLMiddleware(object):
-    def __init__(self, app, xom):
-        self.app = app
-        self.xom = xom
-
-    def __call__(self, environ, start_response):
-        outside_url = environ.get('HTTP_X_OUTSIDE_URL')
-        if not outside_url:
-            outside_url = self.xom.config.args.outside_url
-        if outside_url:
-            # XXX memoize it for later access from replica thread
-            # self.xom.current_outside_url = outside_url
-            outside_url = urlparse(outside_url)
-            environ['wsgi.url_scheme'] = outside_url.scheme
-            environ['HTTP_HOST'] = outside_url.netloc
-            if outside_url.path:
-                environ['SCRIPT_NAME'] = outside_url.path
-        return self.app(environ, start_response)
-
-
 def route_url(self, *args, **kw):
     url = super(self.__class__, self).route_url(*args, **kw)
     # Unquote plus signs in path segment. The settings in pyramid for
